requirements.txt
Django>=4.2
djangorestframework
gunicorn
aiogram
psycopg2-binary
Pillow
drf-yasg